refactor(api): move error to evaluation folder
