+++ api/lib/application/error-manager.js
@@ -37,9 +37,6 @@ function _mapToHttpError(error) {
   if (error instanceof DomainErrors.AccountRecoveryUserAlreadyConfirmEmail) {
     return new HttpErrors.ConflictError(error.message);
   }
-  if (error instanceof DomainErrors.ImproveCompetenceEvaluationForbiddenError) {
-    return new HttpErrors.ImproveCompetenceEvaluationForbiddenError(error.message);
-  }
   if (error instanceof DomainErrors.AlreadyRatedAssessmentError) {
     return new HttpErrors.PreconditionFailedError('Assessment is already rated.');
   }
