+++ api/tests/unit/domain/usecases/improve-competence-evaluation_test.js
@@ -1,5 +1,5 @@
 import { MAX_REACHABLE_LEVEL } from '../../../../lib/domain/constants.js';
-import { ImproveCompetenceEvaluationForbiddenError } from '../../../../lib/domain/errors.js';
+import { ImproveCompetenceEvaluationForbiddenError } from '../../../../src/evaluation/domain/errors.js';
 import { improveCompetenceEvaluation } from '../../../../src/evaluation/domain/usecases/improve-competence-evaluation.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
